% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{getGenerations}
\alias{getGenerations}
\title{Count generations}
\usage{
getGenerations(Ped)
}
\arguments{
\item{Ped}{Dataframe, pedigree with the first three columns being ID - dam -
sire. Column names are ignored, as are additional columns.}
}
\value{
The input pedigree with three columns added: gen, gen.dam, and
  gen.sire
}
\description{
For each individual in a pedigree, count the number of
  generations since the most distant pedigree founder
}
\keyword{internal}
