% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{SimGeno}
\alias{SimGeno}
\title{Simulated genotypes}
\usage{
SimGeno(
  Pedigree,
  nSnp = 400,
  ParMis = 0.4,
  MAF = 0.3,
  CallRate = 0.99,
  SnpError = 5e-04,
  ErrorFM = "version2.0",
  ReturnStats = FALSE,
  OutFile = NA,
  Inherit = "autosomal",
  InheritFile = NA,
  quiet = FALSE,
  PropLQ,
  MisHQ,
  MisLQ,
  ErHQ,
  ErLQ
)
}
\arguments{
\item{Pedigree}{Dataframe, pedigree with the first three columns being id -
dam - sire. Column names are ignored, as are additional columns, with the
exception of a 'Sex' column when Inherit is not 'autosomal'.}

\item{nSnp}{number of SNPs to simulate.}

\item{ParMis}{Single number or vector length two with proportion of parents
with fully missing genotype. Ignored if CallRate is a named vector.}

\item{MAF}{minimum minor allele frequency, and allele frequencies will be
sampled uniformly between this minimum and 0.5, OR a vector with minor
allele frequency at each locus. In both cases, this is the MAF among
pedigree founders, the MAF in the sample will deviate due to drift.}

\item{CallRate}{Either a single number for the mean call rate (genotyping
success), OR a vector with the call rate at each SNP, OR a named vector
with the call rate for each individual. In the third case, ParMis is
ignored, and individuals in the pedigree (as id or parent) not included in
this vector are presumed non-genotyped.}

\item{SnpError}{mean per-locus genotyping error rate across SNPs, and a
beta-distribution will be used to simulate the number of missing cases per
SNP, OR a vector with the genotyping error for each SNP.}

\item{ErrorFM}{function taking the error rate (scalar) as argument and
returning a 3x3 matrix with probabilities that actual genotype i (rows) is
observed as genotype j (columns). Inbuilt ones are as used in sequoia
'version2.0', 'version1.3', or 'version1.1'. See details.}

\item{ReturnStats}{in addition to the genotype matrix, return the input
parameters and mean & quantiles of MAF, error rate and call rates.}

\item{OutFile}{file name for simulated genotypes. If NA (default), return
results within R.}

\item{Inherit}{inheritance pattern, scalar or vector of length nSnp,
Defaults to 'autosomal'. An excel file included in the package has
inheritance patterns for the X and Y chromosome and mtDNA, and allows
custom inheritance patterns. Note that these are NOT currently supported by
the pedigree reconstruction}

\item{InheritFile}{file name for excel file with inheritance patterns,
requires library xlsx.}

\item{quiet}{suppress messages.}

\item{PropLQ}{[deprecated] proportion of low-quality samples.}

\item{MisHQ}{[deprecated] average missingness for high-quality samples,
assuming a beta-distribution with alpha = 1.}

\item{MisLQ}{[deprecated] average missingness in low-quality samples.}

\item{ErHQ}{[deprecated] error rate in high quality samples (defaults to
0.005).}

\item{ErLQ}{[deprecated] error rate in low quality samples.}
}
\value{
if \code{ReturnStats=FALSE} (the default), a matrix with genotype data in
  sequoia's input format, encoded as 0/1/2/-9.

  If \code{ReturnStats=TRUE}, a named list with three elements: list
  'ParamsIN', matrix 'SGeno', and list 'StatsOUT':
  \item{AF}{Frequency in 'observed' genotypes of '1' allele}
  \item{AF.act}{Allele frequency in 'actual' (without genotyping errors &
    missingness)}
  \item{SnpError}{Error rate per SNP (actual /= observed AND observed /=
    missing)}
  \item{SnpCallRate}{Non-missing per SNP}
  \item{IndivError}{Error rate per individual}
  \item{IndivCallRate}{Non-missing per individual}
}
\description{
Simulate SNP genotype data from a pedigree, with optional
  missingess and errors.
}
\details{
There are lots of details.
}
\section{Disclaimer}{
 This simulation is highly simplistic and assumes that
  all SNPs segregate completely independently, that the SNPs are in
  Hardy-Weinberg equilibrium in the pedigree founders. It assumes that
  genotyping errors are not due to heritable mutations of the SNPs, and that
  missingness is random and not e.g. due to heritable mutations of SNP
  flanking regions. Results based on this simulated data will provide an
  minimum estimate of the number of SNPs required, and an optimistic estimate
  of pedigree reconstruction performance.
}

\seealso{
other stuff.
}
\author{
Jisca Huisman, \email{jisca.huisman@gmail.com}
}
